schoolService/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.10
mock==4.0.2
pytz==2020.1
relaxml==0.1.3
requests==2.24.0
simplejson==3.17.0
sqlparse==0.3.1
three==0.8.0
urllib3==1.25.9
